slightly better error handling
the rust backend no longer has to return the same String twice just to report an error. Instead, the JS just checks if it returns true or false (returns the same bool twice now lmao) and if its false it calls get_last_error in the rust backend, which, well, gives you the last error.
im not sure if this change is even worth it, but at least the js frontend isn't checking for a String that says "we gud"

i don't know if anyone can actually read what i just wrote (this commit message, but maybe the code too) but like yeah